JavaScript is Complete
All the JS code is fully documented and has been restructured and organized to be efficient and easy to read.
Many notes were added to the .txt file which serves as a rational for certain decisions and software design philosophies.
the HTML code was reformatted and the buttons were reordered to match a more sensible and logical order, and to reflect the order of the corresponding JS functions